Refactor : ChattingTab의 코드 분리
- 임의로 선언되어있던 type 들을 cam.ts 로 이동시켰습니다.
- Date 객체를 사용하기 좋은 객체로 변환하는 getCurrentDate 메서드를 별도의 파일로 분리하였습니다.
