+++ frontend/src/types/cam.ts
@@ -1,4 +1,5 @@
 import { MediaConnection } from 'peerjs';
+import CurrentDate from './date';
 
 type Status = { video: boolean; audio: boolean; stream: boolean; speaking: boolean };
 type Screen = { userId: string; stream: MediaStream; call: MediaConnection };
@@ -11,4 +12,16 @@ type Control = {
   audio: boolean;
 };
 
-export type { Status, Screen, UserInfo, Control };
+type CamMessageInfo = {
+  msg: string;
+  room: string | null;
+  user: string;
+  date: CurrentDate;
+};
+
+type CamRoomInfo = {
+  socketId: string;
+  userNickname: string;
+};
+
+export type { Status, Screen, UserInfo, Control, CamMessageInfo, CamRoomInfo };
