+++ frontend/src/types/date.ts
@@ -1,0 +1,9 @@
+type CurrentDate = {
+  year: number;
+  month: number;
+  date: number;
+  hour: number;
+  minutes: number;
+};
+
+export default CurrentDate;
