+++ frontend/src/utils/getCurrentDate.ts
@@ -1,0 +1,14 @@
+import CurrentDate from '../types/date';
+
+const getCurrentDate = (): CurrentDate => {
+  const today: Date = new Date();
+  return {
+    year: today.getFullYear(),
+    month: today.getMonth() + 1,
+    date: today.getDate(),
+    hour: today.getHours(),
+    minutes: today.getMinutes(),
+  };
+};
+
+export default getCurrentDate;
